Text and stat section combined
