+++ src/pages/Home.js
@@ -1,18 +1,14 @@
 import React from "react";
 import { Row, Col } from "antd";
-import TextSection from "../components/TextSection";
-import StatisticsSection from "../components/StatisticsSection";
 import TestimonialSlider from "../components/TestimonialSlider";
+import TextAndStats from "../components/TextAndStats";
 
 const Home = () => {
   return (
     <div style={{ minHeight: "100vh", padding: "40px", background: "#f0f2f5" }}>
       <Row gutter={[16, 16]}>
         <Col span={24}>
-          <TextSection />
-        </Col>
-        <Col span={24}>
-          <StatisticsSection />
+          <TextAndStats />
         </Col>
         <Col span={24}>
           <TestimonialSlider />
